# backend/datahub/openapi/datahub-v1.yaml
openapi: 3.1.0
info:
  title: AutoJMS DataHub API
  version: 1.0.0
  description: |
    Phase-1 API contract for the VPS-hosted DataHub. PostgreSQL is private to the
    API deployment and clients use REST plus SignalR. The signed DataHub license
    channel is independent from the desktop update channel.

    `scanTime` is the only business timestamp. A naive JMS value in
    `yyyy-MM-dd HH:mm:ss` is interpreted as `Asia/Ho_Chi_Minh`; an ISO value with
    `Z` or an explicit offset is interpreted using that offset. `uploadTime` is
    accepted only inside the raw `payload` object and is never used for ordering,
    fingerprinting, projection reduction, or retention.
  license:
    name: Internal use
    identifier: internal
servers:
  - url: https://datahub.example.com
    description: Production DataHub (the DNS target can move without changing a license)
  - url: https://datahub-dev.example.com
    description: Staging DataHub (independent database and credentials)
...
paths:
  /api/v1/devices/enroll:
    post:
      tags: [Enrollment]
      operationId: enrollDevice
      summary: Enroll a device from a signed DataHub license assertion
      description: |
        This is the only operation that accepts a signed license assertion. Send it
        as `Authorization: Bearer <signed-license-assertion>`. Production assertions
        are JWS values validated against the configured asymmetric issuer/JWKS.
        The integration-only staging test issuer uses a private
        `v1.<base64url-json>.<HMAC>` format and cannot be enabled for production.
        The API validates signature, expiry, `channel`, and `site_codes` against its
        configured `DATAHUB_CHANNEL` and an existing site. Enrollment never creates a site.
        The response contains a derived device bearer token; the license assertion
        must not be used for any later operation.

        `siteCode` is a requested site selector, not an authorization grant. It must
        be present in the signed assertion's `site_codes`; a body or header channel
        is ignored. `datahub_url` is a signed HTTPS license claim used by the client
        for endpoint selection and is not accepted as request input.
      security:
        - LicenseAssertion: []
      x-license-assertion-only: true
      x-request-channel-authority: signed-license-claim
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnrollRequest'
      responses:
        '201':
          description: Device enrolled and a derived device token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Signed license channel or site scope is not accepted
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                channelMismatch:
                  $ref: '#/components/examples/ChannelMismatch'
                siteNotLicensed:
                  $ref: '#/components/examples/SiteNotLicensed'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Active seat limit reached or device identity conflicts with an existing device
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /api/v1/sites/{siteId}/jms/ingest:
    parameters:
      - $ref: '#/components/parameters/SiteId'
    post:
      tags: [JMS]
      operationId: ingestJmsBulk
      summary: Ingest a bounded bulk JMS page from the current lease leader
      description: |
        This is the bulk dashboard fetch path. It requires a valid device token and
        `X-Leader-Term` matching the site's unexpired lease. The endpoint and
        `/jms/observations` invoke the same canonical `IngestPipeline`; only this
        endpoint performs lease fencing. The server parses `scanTime`, fingerprints
        each observation, reduces all three projection slots, allocates the per-site
        change cursor in the same transaction, and publishes a SignalR doorbell only
        after commit.

        A request is at most 1 MiB and 200 items. The Windows Service chunks a larger
        JMS response and uses a distinct idempotency key for every chunk.
      security:
        - DeviceBearer: []
      x-ingest-pipeline: IngestPipeline
      x-lease-fence: required
      x-max-body-bytes: 1048576
      x-max-items: 200
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/LeaderTerm'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestRequest'
      responses:
        '200':
          description: Chunk accepted, or the stored idempotent response was replayed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResponse'
        '400':
          description: Invalid scan time or malformed request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                invalidScanTime:
                  $ref: '#/components/examples/InvalidScanTime'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Channel or site claim is not accepted
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                channelMismatch:
                  $ref: '#/components/examples/ChannelMismatch'
                siteNotLicensed:
                  $ref: '#/components/examples/SiteNotLicensed'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Lease term or idempotency key conflict
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                leaderFenced:
                  $ref: '#/components/examples/LeaderFenced'
                idempotencyKeyReused:
                  $ref: '#/components/examples/IdempotencyKeyReused'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/sites/{siteId}/jms/observations:
    parameters:
      - $ref: '#/components/parameters/SiteId'
    post:
      tags: [JMS]
      operationId: ingestJmsObservation
      summary: Ingest observations produced by a local interactive operation
      description: |
        This is the interactive path for the Windows Service on the machine where
        the user performed the JMS operation. It uses the same `IngestPipeline` as
        bulk ingest, including fingerprinting, `scanTime` parsing, projection
        reduction, idempotency, and per-site change allocation, but it does not
        require a bulk lease term. The UI never calls JMS directly.

        `X-Leader-Term` is intentionally not part of this operation. Channel and site
        authorization come only from the derived device token and the path claim.
        Body/header channel or site fields are not authority.
      security:
        - DeviceBearer: []
      x-ingest-pipeline: IngestPipeline
      x-lease-fence: none
      x-max-body-bytes: 1048576
      x-max-items: 200
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestRequest'
      responses:
        '200':
          description: Chunk accepted, or the stored idempotent response was replayed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResponse'
        '400':
          description: Invalid scan time or malformed request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                invalidScanTime:
                  $ref: '#/components/examples/InvalidScanTime'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/IdempotencyKeyReused'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    LicenseAssertion:
      type: http
      scheme: bearer
      bearerFormat: signed-license-assertion
      description: |
        Signed DataHub license assertion accepted only by `POST /api/v1/devices/enroll`.
        It carries the required `channel` (`staging` or `production`), `site_codes`,
        expiry, and optionally a signed HTTPS `datahub_url`. It never carries JMS
        tokens, PostgreSQL credentials, DPAPI blobs, or device secrets. Production
        uses asymmetric JWS verification. The custom HMAC assertion is an explicitly
        enabled staging integration seam only; production readiness remains unhealthy
        until the asymmetric verifier is installed.
    DeviceBearer:
      type: http
      scheme: bearer
      bearerFormat: device-token
      description: |
        Derived device credential returned by enrollment. Claims include `device_id`,
        `site_id`, `channel`, role, and `token_version`. It is required for every
        endpoint except health and enrollment. The API rejects a channel claim that
        differs from `DATAHUB_CHANNEL` and never trusts a raw channel/site header or
        body field.
  parameters:
    SiteId:
      name: siteId
      in: path
      required: true
      description: Immutable site UUID. Must equal the authenticated device claim.
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: UUID-like key retained for 24 hours and bound to the SHA-256 request body hash.
      schema:
        type: string
        minLength: 8
        maxLength: 128
    LeaderTerm:
      name: X-Leader-Term
      in: header
      required: true
      description: Current site lease fencing term. Required only on bulk `/jms/ingest`.
      schema:
        type: integer
        format: int64
        minimum: 1
...
  schemas:
    EnrollRequest:
      type: object
      additionalProperties: false
      required: [siteCode, deviceName]
      properties:
        siteCode:
          type: string
          minLength: 1
          maxLength: 64
          description: Existing site requested from the signed assertion's site_codes scope; not an authorization grant.
        deviceName:
          type: string
          minLength: 1
          maxLength: 128
        role:
          type: string
          enum: [operator]
          default: operator
          description: Phase 1 only issues the least-privilege operator role; elevated roles require a later server-side policy.
    EnrollmentResponse:
      type: object
      required: [deviceId, siteId, siteCode, channel, tokenType, deviceToken, tokenVersion, expiresAt]
      properties:
        deviceId:
          type: string
          format: uuid
        siteId:
          type: string
          format: uuid
        siteCode:
          type: string
        channel:
          $ref: '#/components/schemas/DataHubChannel'
        tokenType:
          type: string
          const: Bearer
        deviceToken:
          type: string
          minLength: 1
        tokenVersion:
          type: integer
          minimum: 1
        expiresAt:
          type: string
          format: date-time
    DataHubChannel:
      type: string
      enum: [staging, production]
...
    IngestRequest:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 200
          items:
            $ref: '#/components/schemas/JmsObservation'
    JmsObservation:
      type: object
      additionalProperties: false
      required: [waybillNo, scanTime, payload]
      properties:
        waybillNo:
          type: string
          minLength: 1
          maxLength: 128
        scanTime:
          type: string
          minLength: 19
          description: |
            Required business timestamp. Accept `yyyy-MM-dd HH:mm:ss` (interpreted
            in Asia/Ho_Chi_Minh) or ISO-8601 with Z/explicit offset. Invalid values
            are rejected as INVALID_SCAN_TIME; the server never substitutes now.
          pattern: '^\\d{4}-\\d{2}-\\d{2}(?: \\d{2}:\\d{2}:\\d{2}|T\\d{2}:\\d{2}:\\d{2}(?:\\.\\d+)?(?:Z|[+-]\\d{2}:\\d{2}))$'
        code:
          type: integer
        status:
          type: string
        scanTypeName:
          type: string
        scanNetworkCode:
          type: string
        scanByCode:
          type: string
        packageNumber:
          type: string
        taskCode:
          type: string
        remark1:
          type: string
        remark2:
          type: string
        remark3:
          type: string
        remark4:
          type: string
        remark5:
          type: string
        remark6:
          type: string
        remark7:
          type: string
        remark8:
          type: string
        remark9:
          type: string
        payload:
          type: object
          additionalProperties: true
          description: Raw JMS item for diagnostics. `uploadTime` may appear here only and has no business semantics.
    IngestResponse:
      type: object
      required: [siteId, acceptedItems, duplicateItems, changedProjections, replayed]
      properties:
        siteId:
          type: string
          format: uuid
        acceptedItems:
          type: integer
          minimum: 0
        duplicateItems:
          type: integer
          minimum: 0
        changedProjections:
          type: integer
          minimum: 0
        replayed:
          type: boolean
          description: True when the idempotency record returned a prior response.
        firstChangeSeq:
          type: [integer, 'null']
          format: int64
        lastChangeSeq:
          type: [integer, 'null']
          format: int64
...
    Problem:
      type: object
      required: [type, title, status, code, detail, traceId]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          $ref: '#/components/schemas/ErrorCode'
        detail:
          type: string
        traceId:
          type: string
        instance:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
    ErrorCode:
      type: string
      enum:
        - BAD_REQUEST
        - INVALID_SCAN_TIME
        - UNAUTHORIZED
        - FORBIDDEN
        - CHANNEL_MISMATCH
        - SITE_NOT_LICENSED
        - NOT_FOUND
        - LEADER_FENCED
        - LEASE_HELD
        - SEAT_LIMIT_REACHED
        - DEVICE_CONFLICT
        - IDEMPOTENCY_KEY_REUSED
        - IDEMPOTENCY_IN_PROGRESS
        - RESYNC_REQUIRED
        - PAYLOAD_TOO_LARGE
        - VALIDATION_FAILED
        - RATE_LIMITED
        - SERVICE_UNAVAILABLE
  examples:
    ChannelMismatch:
      summary: License/device channel differs from DATAHUB_CHANNEL
      value:
        type: https://datahub.example.com/problems/channel-mismatch
        title: DataHub channel mismatch
        status: 403
        code: CHANNEL_MISMATCH
        detail: The signed license or device token channel is not accepted by this deployment.
        traceId: 00-00000000000000000000000000000000-0000000000000000-00
    SiteNotLicensed:
      summary: Site is outside signed site_codes
      value:
        type: https://datahub.example.com/problems/site-not-licensed
        title: Site is not licensed
        status: 403
        code: SITE_NOT_LICENSED
        detail: The requested site is not present in the signed license site_codes claim.
        traceId: 00-00000000000000000000000000000000-0000000000000000-00
    InvalidScanTime:
      summary: Scan time cannot be parsed
      value:
        type: https://datahub.example.com/problems/invalid-scan-time
        title: Invalid scanTime
        status: 400
        code: INVALID_SCAN_TIME
        detail: scanTime must be a valid JMS timestamp; no server-local fallback is used.
        traceId: 00-00000000000000000000000000000000-0000000000000000-00
    LeaderFenced:
      summary: Stale or invalid leader term
      value:
        type: https://datahub.example.com/problems/leader-fenced
        title: Leader fenced
        status: 409
        code: LEADER_FENCED
        detail: The device, term, or lease expiry no longer matches the current leader.
        traceId: 00-00000000000000000000000000000000-0000000000000000-00
    IdempotencyKeyReused:
      summary: Key reused with a different body hash
      value:
        type: https://datahub.example.com/problems/idempotency-key-reused
        title: Idempotency key reused
        status: 409
        code: IDEMPOTENCY_KEY_REUSED
        detail: The idempotency key is already bound to a different request body hash.
        traceId: 00-00000000000000000000000000000000-0000000000000000-00
  responses:
    BadRequest:
      description: Request is malformed or contains an invalid business timestamp
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Missing, expired, malformed, or invalid bearer credential
      headers:
        WWW-Authenticate:
          schema:
            type: string
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Authenticated but not authorized for the deployment or site
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Site, device, or requested resource was not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    IdempotencyKeyReused:
      description: Idempotency key is bound to a different body hash
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            idempotencyKeyReused:
              $ref: '#/components/examples/IdempotencyKeyReused'
...
    PayloadTooLarge:
      description: Request exceeds the 1 MiB body or 200-item limit
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ValidationFailed:
      description: JSON schema or domain validation failed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    RateLimited:
      description: Caller exceeded the endpoint rate limit
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ServiceUnavailable:
      description: API or required PostgreSQL dependency is unavailable
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'